update :: 타입 변경

+++ src/main/kotlin/com/ondosee/domain/weather/service/data/req/GetTodayWeatherRequestData.kt
@@ -1,7 +1,7 @@
 package com.ondosee.domain.weather.service.data.req
 
 data class GetTodayWeatherRequestData(
-    val x: Int,
-    val y: Int,
+    val x: Long,
+    val y: Long,
     val location: String
 )
